Build PPI Network

Uses all inputted database information on protein-protein interactions in order to create a combined PPI-network. All edges in the PPI network are checked against the cocluster correlation network to ensure existence and a non-zero weight. If either of these conditions are not met, then it will be removed from the list of PPI edges. This new, cluster filtered network is then assigned to the global namespace.

Usage

BuildClusterFilteredNetwork(
  stringdb.edges,
  genemania.edges,
  kinsub.edges,
  gene.cccn.edges,
  db.filepaths = c()
)

Arguments

stringdb.edges

Data frame of consisting of the network of interactions from the genes of study pulled from the STRINGdb database

genemania.edges

GeneMANIA network of protein-protein interactions from the genes of study; defaults to NA

kinsub.edges

TODO

gene.cccn.edges

TODO

db.filepaths

A vector of paths to the additional ppi network files; defaults to an empty vector

Value

A version of ppi.network with only the edges that exist in gene.cccn and have non-zero weights

Examples

ClusterFilteredNetwork(ex.gene.cccn, ex.ppi.network, cfn.name = "ex.cfn")
#> Error in ClusterFilteredNetwork(ex.gene.cccn, ex.ppi.network, cfn.name = "ex.cfn"): could not find function "ClusterFilteredNetwork"
utils::head(ex.cfn)
#>    Gene.1 Gene.2                                     Interaction
#> 1    MYH9 PIK3R2 experimental_transferred, Physical Interactions
#> 2    MYH9 ABLIM1             experimental, Physical Interactions
#> 3   OXCT1 ZDHHC5                                    experimental
#> 4 SUPT16H ZDHHC5                                    experimental
#> 5   USP15 ZDHHC5                                    experimental
#> 6    MYH9 ATP5PO                                    experimental
#>          PPI.weight
#> 1 0.215459870486844
#> 2 0.120127317943767
#> 3 0.209019327129563
#> 4 0.209019327129563
#> 5 0.209019327129563
#> 6 0.137437365783822