Build PPI Network

Uses all inputted database information on protein-protein interactions in order to create a combined PPI-network.

Usage

BuildPPINetwork(
  stringdb.edges = NA,
  gm.network = NA,
  db.filepaths = c(),
  ppi.network.name = "ppi.network"
)

Arguments

stringdb.edges

Data frame of consisting of the network of interactions from the genes of study pulled from the STRINGdb database

gm.network

GeneMANIA network of protein-protein interactions from the genes of study; defaults to NA

db.filepaths

A vector of paths to the additional ppi network files; defaults to an empty vector

ppi.network.name

Desired name for the output protein-protein interaction network using all entered database input; defaults to ppi.network

Value

A dataframe representing how strongly proteins are known to interact based on previous research contained within provided database networks

Examples

ex.string <- ex.stringdb.edges
ex.gm <- ex.gm.network
BuildPPINetwork(stringdb.edges = ex.string, gm.network = ex.gm, ppi.network.name = "ex.ppi.network")
utils::head(ex.ppi.network)
#>     Gene.1 Gene.2                                     Interaction
#> 1     MYH9 PIK3R2 experimental_transferred, Physical Interactions
#> 11    MYH9 ABLIM1             experimental, Physical Interactions
#> 25   OXCT1 ZDHHC5                                    experimental
#> 27 SUPT16H ZDHHC5                                    experimental
#> 33   USP15 ZDHHC5                                    experimental
#> 35    MYH9 ATP5PO                                    experimental
#>    STRINGdb.combined_score   GM.weights
#> 1                0.1546170 0.0608428340
#> 11               0.1195419 0.0005854425
#> 25               0.2090193           NA
#> 27               0.2090193           NA
#> 33               0.2090193           NA
#> 35               0.1374374           NA