Filter PPI Network by Known Clusters

All edges in the PPI network are checked against the cocluster correlation network to ensure existence and a non-zero weight. If either of these conditions are not met, then it will be removed from the list of PPI edges. This new, cluster filtered network is then assigned to the global namespace.

Usage

ClusterFilteredNetwork(gene.cccn, ppi.network, cfn.name = "cfn")

Arguments

gene.cccn

A matrix showing strength of relationships between proteins using the common clusters between the three distance metrics (Euclidean, Spearman, and Combined (SED))

ppi.network

A dataframe representing how strongly proteins are known to interact based on previous research contained within provided database networks

cfn.name

Desired name of the output cluster filtered network

Value

A version of ppi.network with only the edges that exist in gene.cccn and have non-zero weights

Examples

ClusterFilteredNetwork(ex.gene.cccn, ex.ppi.network, cfn.name = "ex.cfn")
#> Error: object 'ex.gene.cccn' not found
utils::head(ex.cfn)
#>    Gene.1 Gene.2                         Interaction         PPI.weight
#> 1    MYH9 ABLIM1 experimental, Physical Interactions  0.124660897834244
#> 2    MYH9 ATP5PO                        experimental  0.142644873699851
#> 3    MYH9 SEC16A experimental, Physical Interactions 0.0922490307565398
#> 4   RBM14 SEC16A experimental, Physical Interactions  0.177836973750756
#> 5  TUBA1B EEF1A1                        experimental  0.269687964338782
#> 6 PLEKHA6  PARD3                        experimental   0.12332838038633