Get STRINGdb PPI data

This function finds protein-protein interaction weights by consulting utilizing the STRINGdb database.

Usage

GetSTRINGdb(
  cccn.matrix,
  stringdb.name = "stringdb.edges",
  nodenames.name = "nodenames"
)

Arguments

cccn.matrix

A matrix showing strength of relationships between proteins using common clusters between the three distance metrics (Euclidean, Spearman, and Combined (SED))

stringdb.name

Desired name for the output STRINGdb data frame; defaults to "stringdb.edges"

nodenames.name

Desired name for list of gene names; defaults to nodenames

Value

Data frame of consisting of the network of interactions from the genes of study pulled from the STRINGdb database and a list of gene names

Examples

GetSTRINGdb(ex.cccn.matrix, 'ex.stringdb.edges', 'ex.nodenames')
#>   Gene.Names            STRING_id
#> 1     ABLIM1 9606.ENSP00000277895
#> 2       PKP4 9606.ENSP00000374409
#> 3       CTTN 9606.ENSP00000365745
#> 4       MYH9 9606.ENSP00000216181
#> 5      PARD3 9606.ENSP00000363921
#> 6    TMEM192 9606.ENSP00000305069
utils::head(ex.stringdb.edges)
#>     Gene.1   Gene.2              Interaction STRINGdb.combined_score
#> 9     MYH9   ABLIM1             experimental                     167
#> 23    MYH9   ATP5PO             experimental                     192
#> 25 SLC38A2   ATP5PO             experimental                     221
#> 29     APP   ATP5PO experimental_transferred                     337
#> 35  ATP5PO ATP6V0D1                 database                     744
#> 59    ATG3    HSPA9 experimental_transferred                     218
utils::head(ex.nodenames)
#>   Gene.Names
#> 1     ABLIM1
#> 2       PKP4
#> 3       CTTN
#> 4       MYH9
#> 5      PARD3
#> 6    TMEM192