Graph Cluster Filtered Network

Creates a cytoscape graph of the cluster filtered network. Ensure that you have the cytoscape app open and the RCy3 package downloaded and libraried.

GraphCfn(
  cfn,
  ptmtable,
  funckey = PTMsToPathways::ex.funckey,
  Network.title = "cfn",
  Network.collection = "PTMsToPathways",
  visual.style.name = "PTMsToPathways.style",
  background.color = "#949494",
  edge.label.color = "#17202a",
  node.label.color = "#000000",
  default.font = "Times New Roman",
  node.font.size = 20,
  edge.font.size = 8,
  edge.line.style = "SOLID",
  edge.opacity = 175,
  edge.label.opacity = 255,
  border.opacity = 255,
  node.label.opacity = 255,
  node.fill.opacity = 255
)

Arguments

cfn

A version of ppi.network with only the edges that exist in cccn.matrix and have non-zero weights

ptmtable

A dataset for post-translational modifications. Formatted with numbered rows, and the first column containing PTM names. The rest of the column names should be drugs. Values are numeric values that represent how much the PTM has reacted to the drug.

funckey

A table graphing gene names to type of protein; defaults to internal database at PTMsToPathways::ex.funckey

Network.title

Desired title for the created Cytoscape Network; defaults to "cfn"

Network.collection

Desired name for the collection created on Cytoscape in which the network will reside; defaults to "PTMsToPathways"

visual.style.name

Desired name for the visual style created on Cytoscape; defaults to "PTMsToPathways.style"

background.color

Hex code of background color of graph; defaults to '#fcf3cf'

edge.label.color

Hex code of edge label color of graph; defaults to '#17202a'

node.label.color

Hex code of node label color of graph; defaults to '#145a32'

default.font

Font style of edge and node names; defaults to "Times New Roman"

node.font.size

Font size of the node name; defaults to 12

edge.font.size

Font size of the edge name; defaults to 8

edge.line.style

Type of edge style; defaults to "SOLID"; options include: "PARALLEL_LINES", "MARQUEE_EQUAL", "DOT", "EQUAL_DASH", "LONG_DASH", "CONTIGUOUS_ARROW", "MARQUEE_DASH", "DASH_DOT", "BACKWARD_SLASH", "FORWARD_SLASH", "VERTICAL_SLASH", "SOLID", "SEPARATE_ARROW", "MARQUEE_DASH_DOT", "ZIGZAG", "SINEWAVE"

edge.opacity

Opacity of the edge line on a scale of 0 - 255 with 0 being transparent; defaults to 175

edge.label.opacity

Opacity of the edge label on a scale of 0 - 255 with 0 being transparent; defaults to 255

border.opacity

Opacity of the node border on a scale of 0 - 255 with 0 being transparent; defaults to 255

node.label.opacity

Opacity of the node label on a scale of 0 - 255 with 0 being transparent; defaults to 255

node.fill.opacity

Opacity of the node fill on a scale of 0 - 255 with 0 being transparent; defaults to 255

Value

A cytoscape graph of the cluster filtered network

Details

If error occurs, run cytoscapePing() to ensure connection to the Cytoscape Interface and try again.

Examples

# GraphCFN(ex.cfn)
# See vignette for default graph