Process GM Edge File

This function processes the GM edgefile and translates it back into gene names using the nodetable.

Usage

ProcessGMEdgefile(
  gm.edgefile.path,
  gm.nodetable.path,
  db_nodes.path,
  gm.network.name = "gm.network",
  returndata = TRUE
)

Arguments

gm.edgefile.path

Path to GeneMANIA edgefile

gm.nodetable.path

Path to GeneMANIA nodetable

db_nodes.path

Path to the node file from MakeDBInput

gm.network.name

Desired name for the output genemania network; defaults to gm.network

Value

Data frame of consisting of the network of interactions from the genes of study

Examples

ex.edgefile <- system.file("extdata/ex_gm_edgetable.csv", package = "PTMsToPathways")
ex.nodefile <- system.file("extdata/ex_gm_nodetable.csv", package = "PTMsToPathways")
ex.db.nodes  <- system.file("extdata/ex_db_nodes.txt", package = "PTMsToPathways")
ProcessGMEdgefile(ex.edgefile, ex.nodefile, ex.db.nodes, "ex.gm.network")
#> Error in start_line[1]:(end_line[1] - 1): NA/NaN argument
utils::head(ex.gm.network)
#>    Gene.1  Gene.2 interaction   GM.weights
#> 1     APP   APH1A     Pathway 6.132073e-03
#> 6   SF3B2 HNRNPH2     Pathway 5.446725e-05
#> 7  HNRNPF HNRNPH2     Pathway 5.446725e-05
#> 8  HNRNPF   SF3B2     Pathway 5.446725e-05
#> 9  HNRNPU HNRNPH2     Pathway 5.446725e-05
#> 10 HNRNPU   SF3B2     Pathway 5.446725e-05