Clusters List Example

List of lists of dataframes containing PTMs that have been found to cluster based on their distance in 3D space for a specific distance metric

Usage

ex.clusters.list

Format

List of 3 Distance metrics (Euclidean, Spearman, and the average of Spearman and Euclidean (SED)) and the clusters that form when they are run through tsne:

Euclidean

A list of 54 clusters of ptms based on their Euclidean distance and positional distance determined by tsne

Spearman

A list of 125 clusters of ptms based on their Spearman dissimilarity and positional distance determined by tsne

SED

A list of 39 clusters of ptms based on the average of their Euclidean distance and Spearman Dissimilarity and positional distance determined by tsne

Source

"created by MakeClusterList"

Examples

head(ex.clusters.list)
#> $Euclidean
#> $`1`
#>            PTMnames group
#> 1     ABLIM1 p Y357     1
#> 2     ABLIM1 p Y373     1
#> 3      AGFG1 p Y327     1
#> 4   ATP5PO ack K162     1
#> 5       CTTN p T401     1
#> 6   HNRNPF ack K167     1
#> 7   HNRNPU ubi K524     1
#> 8        LCK p Y394     1
#> 9       MET p Y1003     1
#> 10      MET p Y1235     1
#> 11    MET ubi K1259     1
#> 12       MYH9 p Y11     1
#> 13     ORAI1 p Y300     1
#> 14    PARD3 p Y1127     1
#> 15    PIK3R2 p Y464     1
#> 16     PKP4 p Y1156     1
#> 17      PKP4 p Y470     1
#> 18      PKP4 p Y478     1
#> 19     PTPRK p Y858     1
#> 20    RASSF8 p Y418     1
#> 21    SCRIB p Y1268     1
#> 22 SUPT16H ack K674     1
#> 23   TMEM192 p Y213     1
#> 24   TUBA1B p Y103;     1
#> 25   TUBA1B p Y272;     1
#> 26 TUBA1B ubi K112;     1
#> 27 TUBA1B ubi K326;     1
#> 28 TUBA1B ubi K336;     1
#> 29 TUBA1B ubi K338;     1
#> 30 TUBA1B ubi K370;     1
#> 31  TUBA1B ubi K60;     1
#> 32  TUBA1B ubi K96;     1
#> 33    ZDHHC5 p S529     1
#> 34    ZDHHC5 p Y533     1
#> 
#> $`2`
#>            PTMnames group
#> 1     ABLIM1 p Y383     2
#> 2     ABLIM1 p Y410     2
#> 3    CTNNAL1 p S538     2
#> 4     DCBLD1 p Y540     2
#> 5     EZR ubi K237;     2
#> 6     MET ubi K1232     2
#> 7     MYH9 ubi K299     2
#> 8     PHLDB2 p Y567     2
#> 9    SEC16A p Y2193     2
#> 10 TUBA1B ubi K311;     2
#> 11    ZDHHC5 p Y497     2
#> 
#> $`3`
#>           PTMnames group
#> 1    ABLIM1 p Y406     3
#> 2  ATP5PO ack K199     3
#> 3  HNRNPU ack K343     3
#> 4  HNRNPU ack K352     3
#> 5    HNRNPU p Y473     3
#> 6      MET p Y1230     3
#> 7   MYH9 ack K1387     3
#> 8  PLEKHA6 p Y1044     3
#> 9 TUBA1B ack K370;     3
#> 
#> $`4`
#>        PTMnames group
#> 1 ABLIM1 p Y461     4
#> 2 HNRNPF p Y266     4
#> 3 PHLDB2 p Y301     4
#> 
#> $`5`
#>            PTMnames group
#> 1     ABLIM1 p Y688     5
#> 2   ABLIM1 ubi K275     5
#> 3   EEF1A1 ubi K395     5
#> 4       EZR ubi K60     5
#> 5   HNRNPU ubi K433     5
#> 6   HSPA1B ubi K597     5
#> 7     MYH9 ack K882     5
#> 8     MYH9 ubi K186     5
#> 9     MYH9 ubi K373     5
#> 10   MYH9 ubi K576;     5
#> 11  PI4K2A ubi K268     5
#> 12   PSMC1 ubi K293     5
#> 13   PSMC1 ubi K396     5
#> 14    PSMC1 ubi K69     5
#> 15    RBM4 ubi K223     5
#> 16   SF3B2 ubi K387     5
#> 17   SF3B2 ubi K394     5
#> 18  SLC12A4 ubi K97     5
#> 19 SLC12A4 ubi K979     5
#> 20    SPRED1 p Y211     5
#> 
#> $`6`
#>            PTMnames group
#> 1     ABLIM1 ubi K8     6
#> 2      ATP5PO p Y35     6
#> 3  CTNNA2 ack K913;     6
#> 4        CTTN p Y84     6
#> 5     DCBLD1 p Y578     6
#> 6       EZR ack K60     6
#> 7       FRS2 p Y392     6
#> 8     HNRNPU p Y257     6
#> 9     HNRNPU p Y260     6
#> 10  HSPA1B ack K348     6
#> 11    HSPA1B p Y41;     6
#> 12     MERTK p Y549     6
#> 13        OAT p Y69     6
#> 14     PARD3 p Y388     6
#> 15      PKP4 p S406     6
#> 16   PLEKHA6 p S472     6
#> 17   PLEKHA6 p S478     6
#> 18     PTPRH p Y850     6
#> 19     RBM14 p Y249     6
#> 20    SEC16A p Y489     6
#> 
#> $`7`
#>          PTMnames group
#> 1  ACOX1 ack K437     7
#> 2   MYH9 ack K651     7
#> 3  MYH9 ack K821;     7
#> 4   MYH9 ubi K760     7
#> 5  MYH9 ubi K821;     7
#> 6 SLC9A1 ubi K690     7
#> 
#> $`8`
#>            PTMnames group
#> 1      ACOX1 p Y629     8
#> 2      ACTR5 p Y564     8
#> 3   ATP5PO ack K158     8
#> 4     CTTN ack K161     8
#> 5     CTTN ack K171     8
#> 6      DNM2 p Y125;     8
#> 7   EEF1A1 ack K41;     8
#> 8       EMC7 p S114     8
#> 9       EMC7 p T113     8
#> 10        EZR p Y85     8
#> 11   HNRNPH2 p Y306     8
#> 12  HNRNPU ack K265     8
#> 13 HSPA1B ack K319;     8
#> 14     HSPA9 p Y161     8
#> 15       LCK p Y505     8
#> 16    MYO18A p Y700     8
#> 17       NFS1 p Y80     8
#> 18       OAT p Y126     8
#> 19   OXCT1 ack K418     8
#> 20       PHB2 p Y77     8
#> 21    PI4K2A p Y402     8
#> 22    PTPRH p Y1094     8
#> 23  RASSF8 ubi K176     8
#> 24      RMC1 p Y610     8
#> 25   RNASEH2B p S21     8
#> 26   RNASEH2B p Y23     8
#> 27   RPRD2 ack K671     8
#> 28   SF3B2 ack K275     8
#> 29   SUPT16H p Y535     8
#> 
#> $`9`
#>        PTMnames group
#> 1  APH1A p Y256     9
#> 2   CTTN p Y215     9
#> 3 NSFL1C p Y167     9
#> 4  PKP4 p T1174     9
#> 5 SCRIB p Y1360     9
#> 6   SGMS2 p Y56     9
#> 7   TJP3 p Y364     9
#> 8 ZNF185 p Y408     9
#> 
#> $`10`
#>            PTMnames group
#> 1     APH1A ubi K95    10
#> 2     APH1A ubi K96    10
#> 3    ATP5PO ack K60    10
#> 4     CTTN ack K198    10
#> 5     CTTN ubi K230    10
#> 6   EEF1A1 ubi K179    10
#> 7   EEF1A1 ubi K392    10
#> 8      EZR ubi K344    10
#> 9      EZR ubi K72;    10
#> 10 HSPA1B ubi K507;    10
#> 11      LCK p Y394;    10
#> 12   MYH9 ack K1525    10
#> 13   MYH9 ack K1527    10
#> 14   MYH9 ubi K821;    10
#> 15  NSFL1C ubi K127    10
#> 16  PI4K2A ubi K240    10
#> 17    PKP4 ubi K518    10
#> 18     PTPRK p Y916    10
#> 19    SEC16A p Y813    10
#> 20   SF3B2 ubi K280    10
#> 21   SF3B2 ubi K543    10
#> 22   TUBA1B p Y103;    10
#> 23 TUBA1B ubi K336;    10
#> 24 TUBA1B ubi K338;    10
#> 25 TUBA1B ubi K370;    10
#> 26    WEE1 ack K596    10
#> 
#> $`11`
#>           PTMnames group
#> 1       APP p T761    11
#> 2       ATG3 p Y18    11
#> 3  ATP6V0D1 p Y347    11
#> 4   CTNNAL1 p Y534    11
#> 5   CTNNAL1 p Y536    11
#> 6    DCBLD1 p Y589    11
#> 7    DCBLD1 p Y600    11
#> 8     FKBP4 p Y411    11
#> 9      FRS2 p Y349    11
#> 10   HNRNPF p Y210    11
#> 11    LARP1 p Y361    11
#> 12    MERTK p Y929    11
#> 13   PIK3R2 p Y460    11
#> 14     PKP4 p Y390    11
#> 15  PLEKHA6 p Y471    11
#> 16  PLEKHA6 p Y498    11
#> 17    PRRC2A p Y17    11
#> 18    PSMC1 p Y439    11
#> 19    RBM14 p Y226    11
#> 20    RBM14 p Y648    11
#> 21     RBM4 p Y194    11
#> 22  SEC16A p Y1209    11
#> 23     SGMS2 p Y59    11
#> 24   SLC38A2 p Y28    11
#> 25   SLC9A1 p S693    11
#> 26   SLC9A1 p Y683    11
#> 27     SMAGP p Y95    11
#> 28     SYAP1 p Y60    11
#> 29  TMEM106B p Y50    11
#> 30     TNK2 p Y860    11
#> 31  TUBA1B p Y282;    11
#> 32     WEE1 p S139    11
#> 33     WEE1 p Y132    11
#> 34   ZDHHC5 p Y630    11
#> 
#> $`12`
#>          PTMnames group
#> 1      APP p Y757    12
#> 2 EEF1A1 ack K392    12
#> 3     EZR ack K83    12
#> 4     PHB2 p Y248    12
#> 5     PKP4 p Y114    12
#> 
#> $`13`
#>           PTMnames group
#> 1     APP ubi K751    13
#> 2    CTTN ack K107    13
#> 3  EEF1A1 ubi K172    13
#> 4  EEF1A1 ubi K180    13
#> 5  EEF1A1 ubi K439    13
#> 6  EEF1A1 ubi K44;    13
#> 7     EZR ubi K139    13
#> 8  HNRNPF ubi K224    13
#> 9  HNRNPU ack K814    13
#> 10 HSPA1B ubi K524    13
#> 11  HSPA9 ack K135    13
#> 12  HSPA9 ack K600    13
#> 13    MET ubi K962    13
#> 14   MYH9 ubi K860    13
#> 15 NSFL1C ack K172    13
#> 16 SLC38A2 ubi K59    13
#> 
#> $`14`
#>          PTMnames group
#> 1    APP ubi K763    14
#> 2 SLC9A1 ubi K583    14
#> 
#> $`15`
#>           PTMnames group
#> 1     ARAP2 p Y334    15
#> 2  EEF1A1 ack K395    15
#> 3   HSPA9 ack K300    15
#> 4   MYH9 ack K1014    15
#> 5    MYH9 ack K299    15
#> 6    MYH9 ack K545    15
#> 7   MYH9 ack K565;    15
#> 8    MYH9 ack K613    15
#> 9   MYH9 ack K682;    15
#> 10  OXCT1 ack K473    15
#> 11  PSMC1 ack K258    15
#> 
#> $`16`
#>            PTMnames group
#> 1      ARAP2 p Y473    16
#> 2     CTTN ack K124    16
#> 3     CTTN ack K235    16
#> 4     CTTN ack K272    16
#> 5     CTTN ack K309    16
#> 6      CTTN ack K87    16
#> 7       CTTN p S156    16
#> 8       CTTN p S418    16
#> 9       CTTN p Y154    16
#> 10      CTTN p Y162    16
#> 11      CTTN p Y178    16
#> 12      CTTN p Y228    16
#> 13      CTTN p Y265    16
#> 14      CTTN p Y334    16
#> 15      CTTN p Y421    16
#> 16      CTTN p Y446    16
#> 17  EEF1A1 ack K172    16
#> 18  EEF1A1 ack K179    16
#> 19 EEF1A1 ack K255;    16
#> 20  EEF1A1 ack K44;    16
#> 21   EEF1A1 p Y141;    16
#> 22    EEF1A1 p Y177    16
#> 23      EZR p Y270;    16
#> 24     EZR ubi K79;    16
#> 25      FRS2 p Y306    16
#> 26  HNRNPU ack K565    16
#> 27  HNRNPU ack K635    16
#> 28 HSPA1B ack K500;    16
#> 29       LCK p Y192    16
#> 30      MET p S1236    16
#> 31      MET p Y1234    16
#> 32     MYH9 p Y1805    16
#> 33    MYH9 ubi K833    16
#> 34  NSFL1C ubi K172    16
#> 35    PARD3 p Y1080    16
#> 36     PARD3 p Y719    16
#> 37      PKP4 p T477    16
#> 38     PKP4 p Y1094    16
#> 39     PKP4 p Y1115    16
#> 40     PKP4 p Y1168    16
#> 41      PKP4 p Y372    16
#> 42      PKP4 p Y415    16
#> 43      PKP4 p Y420    16
#> 44      PKP4 p Y421    16
#> 45      PKP4 p Y465    16
#> 46      PKP4 p Y494    16
#> 47   PLEKHA6 p Y404    16
#> 48   PLEKHA6 p Y492    16
#> 49    PTPN11 p Y546    16
#> 50    PTPN11 p Y584    16
#> 51     PTPN11 p Y62    16
#> 52    PTPRH p Y1102    16
#> 53   PTTG1IP p Y174    16
#> 54   RBM14 ack K149    16
#> 55 TMEM106B ubi K14    16
#> 56      TNK2 p Y827    16
#> 57      TNK2 p Y859    16
#> 58    ZNF185 p Y581    16
#> 
#> $`17`
#>           PTMnames group
#> 1      ARAP2 p Y77    17
#> 2     NSFL1C p Y95    17
#> 3      PKP4 p Y369    17
#> 4   SEC16A p Y1291    17
#> 5    SPRED1 p Y292    17
#> 6 TUBA1B ubi K304;    17
#> 7     USP15 p Y274    17
#> 
#> $`18`
#>          PTMnames group
#> 1 ATP5PO ack K192    18
#> 2 HSPA1B ack K246    18
#> 3  MYH9 ubi K682;    18
#> 
#> $`19`
#>            PTMnames group
#> 1 ATP6V0D1 ubi K343    19
#> 2     HNRNPU p S271    19
#> 3     MYH9 ack K355    19
#> 
#> $`20`
#>         PTMnames group
#> 1 CTNNA2 p Y417;    20
#> 2      MYH9 p Y9    20
#> 3 PTTG1IP p Y165    20
#> 
#> $`21`
#>       PTMnames group
#> 1  CTTN p S405    21
#> 2 PTPRK p S856    21
#> 
#> $`22`
#>         PTMnames group
#> 1    CTTN p S417    22
#> 2 MYH9 ubi K1415    22
#> 
#> $`23`
#>      PTMnames group
#> 1 CTTN p S447    23
#> 2 TNK2 p Y284    23
#> 
#> $`24`
#>           PTMnames group
#> 1      CTTN p T440    24
#> 2 HSPA1B ubi K423;    24
#> 
#> $`25`
#>        PTMnames group
#> 1   CTTN p Y104    25
#> 2   CTTN p Y199    25
#> 3 EEF1A1 p Y29;    25
#> 4  MYH9 p Y1408    25
#> 
#> $`26`
#>         PTMnames group
#> 1    CTTN p Y141    26
#> 2    MYH9 p Y754    26
#> 3   PARD3 p Y489    26
#> 4   PKP4 p Y1172    26
#> 5    PKP4 p Y425    26
#> 6 RBM14 ack K164    26
#> 
#> $`27`
#>           PTMnames group
#> 1      CTTN p Y289    27
#> 2  HNRNPU ubi K186    27
#> 3     PARD3 p S717    27
#> 4     PKP4 p Y1132    27
#> 5     PTPRK p Y849    27
#> 6 PTTG1IP ubi K164    27
#> 7      TNK2 p T857    27
#> 
#> $`28`
#>           PTMnames group
#> 1      CTTN p Y453    28
#> 2 TMEM192 ubi K254    28
#> 
#> $`29`
#>        PTMnames group
#> 1 CTTN ubi K181    29
#> 2 CTTN ubi K272    29
#> 
#> $`30`
#>            PTMnames group
#> 1     CTTN ubi K193    30
#> 2     DCBLD1 p S619    30
#> 3      EZR ubi K450    30
#> 4      EZR ubi K458    30
#> 5    MYH9 ack K1793    30
#> 6     PIK3R2 p Y505    30
#> 7       PKP4 p S273    30
#> 8       PKP4 p Y306    30
#> 9     PTPN11 p T547    30
#> 10     RBM14 p Y645    30
#> 11   SEC16A p Y1031    30
#> 12 TUBA1B ubi K163;    30
#> 13 TUBA1B ubi K164;    30
#> 14 TUBA1B ubi K394;    30
#> 15 TUBA1B ubi K401;    30
#> 
#> $`31`
#>           PTMnames group
#> 1      DNM2 p Y231    31
#> 2     EZR ubi K143    31
#> 3 HSPA1B ubi K512;    31
#> 4      TNK2 p T517    31
#> 5 TUBA1B ack K336;    31
#> 6   TUBA1B p Y312;    31
#> 
#> $`32`
#>           PTMnames group
#> 1 EEF1A1 ack K146;    32
#> 2 EEF1A1 ubi K146;    32
#> 3  HSPA1B ubi K539    32
#> 4 PTTG1IP ubi K169    32
#> 
#> $`33`
#>           PTMnames group
#> 1  EEF1A1 ack K408    33
#> 2      EZR ubi K83    33
#> 3    HNRNPF p Y253    33
#> 4   MERTK ubi K856    33
#> 5  PI4K2A ubi K336    33
#> 6      PKP4 p Y157    33
#> 7      PKP4 p Y224    33
#> 8    PTPN11 p S595    33
#> 9    SLC12A4 p Y17    33
#> 10   SYAP1 ack K70    33
#> 11     TJP3 p Y361    33
#> 12  TUBA1B p Y357;    33
#> 13    USP15 p Y263    33
#> 14  ZDHHC5 p Y711;    33
#> 
#> $`34`
#>            PTMnames group
#> 1   EEF1A1 ack K439    34
#> 2    MYH9 ack K1828    34
#> 3 RNASEH2B ack K284    34
#> 
#> $`35`
#>         PTMnames group
#> 1 EEF1A1 p T142;    35
#> 2    LCK p T395;    35
#> 3  PARD3 p Y1177    35
#> 4 PRRC2A ack K35    35
#> 
#> $`36`
#>             PTMnames group
#> 1    EEF1A1 ubi K212    36
#> 2       EZR ubi K162    36
#> 3    HNRNPF ubi K299    36
#> 4    HNRNPH2 ubi K35    36
#> 5   HSPA1B ubi K451;    36
#> 6   HSPA1B ubi K507;    36
#> 7      MET ack K1248    36
#> 8        MET p Y1349    36
#> 9        MET p Y1356    36
#> 10    MYH9 ack K1492    36
#> 11    MYH9 ack K1538    36
#> 12     MYH9 ubi K82;    36
#> 13 SLC12A4 ubi K1010    36
#> 14   SLC38A2 ubi K38    36
#> 15  SUPT16H ack K904    36
#> 16  SUPT16H ubi K904    36
#> 17  TUBA1B ack K280;    36
#> 18  TUBA1B ack K304;    36
#> 19  TUBA1B ubi K163;    36
#> 
#> $`37`
#>            PTMnames group
#> 1   EEF1A1 ubi K273    37
#> 2   HNRNPF ubi K167    37
#> 3   HNRNPU ubi K464    37
#> 4     MET ubi K1161    37
#> 5      PARD3 p Y933    37
#> 6    PKP4 ubi K518;    37
#> 7    SF3B2 ubi K529    37
#> 8   SLC38A2 ubi K33    37
#> 9     SPRED1 p Y294    37
#> 10 TUBA1B ubi K352;    37
#> 
#> $`38`
#>          PTMnames group
#> 1 EEF1A1 ubi K408    38
#> 2 HSPA1B ubi K126    38
#> 3     PKP4 p S281    38
#> 4  TMEM106B p Y18    38
#> 
#> $`39`
#>          PTMnames group
#> 1    EZR ubi K209    39
#> 2    EZR ubi K211    39
#> 3  MYH9 ubi K1638    39
#> 4 TUBA1B ubi K96;    39
#> 
#> $`40`
#>            PTMnames group
#> 1      EZR ubi K35;    40
#> 2    HNRNPF ubi K98    40
#> 3   HNRNPU ubi K265    40
#> 4      MYH9 ubi K14    40
#> 5     MYH9 ubi K850    40
#> 6     PSMC1 ubi K91    40
#> 7   PTPRK ubi K1360    40
#> 8     SF3B2 ack K10    40
#> 9  SLC12A4 ubi K988    40
#> 10  SLC9A1 ubi K612    40
#> 11  SLC9A1 ubi K625    40
#> 12  SLC9A1 ubi K767    40
#> 
#> $`41`
#>           PTMnames group
#> 1   FKBP4 ack K354    41
#> 2      MYH9 ack K8    41
#> 3     PARD3 p S728    41
#> 4    PHLDB2 p Y162    41
#> 5   PSMC1 ubi K430    41
#> 6 PTTG1IP ack K169    41
#> 
#> $`42`
#>           PTMnames group
#> 1    FKBP4 ubi K28    42
#> 2      MET p Y1365    42
#> 3 TUBA1B ack K311;    42
#> 
#> $`43`
#>         PTMnames group
#> 1 HNRNPF ack K98    43
#> 2 HSPA9 ack K159    43
#> 3     MET p S990    43
#> 
#> $`44`
#>         PTMnames group
#> 1 HNRNPH2 p Y276    44
#> 2 HSPA9 ubi K121    44
#> 3 SEC16A p Y1380    44
#> 
#> $`45`
#>          PTMnames group
#> 1 HNRNPU ack K464    45
#> 2   MET ubi K1318    45
#> 
#> $`46`
#>          PTMnames group
#> 1 HNRNPU ubi K352    46
#> 
#> $`47`
#>           PTMnames group
#> 1  HNRNPU ubi K635    47
#> 2  HSPA1B ubi K325    47
#> 3  HSPA1B ubi K550    47
#> 4   HSPA1B ubi K56    47
#> 5    MYH9 ubi K102    47
#> 6   PSMC1 ubi K237    47
#> 7  SPRED1 ubi K284    47
#> 8 SUPT16H ubi K596    47
#> 
#> $`48`
#>           PTMnames group
#> 1  HSPA1B ubi K628    48
#> 2 TUBA1B ack K336;    48
#> 
#> $`49`
#>        PTMnames group
#> 1 MET ubi K1190    49
#> 2  PDZD11 p Y36    49
#> 
#> $`50`
#>         PTMnames group
#> 1 MYH9 ack K1802    50
#> 2   MYH9 ack K74    50
#> 3  MYH9 ubi K856    50
#> 4    PKP4 p Y443    50
#> 
#> $`51`
#>       PTMnames group
#> 1 MYH9 ack K47    51
#> 
#> $`52`
#>        PTMnames group
#> 1 MYH9 ubi K651    52
#> 2   PKP4 p Y275    52
#> 3 PTPN11 p Y279    52
#> 4 SLC38A2 p Y41    52
#> 
#> $`53`
#>          PTMnames group
#> 1 NSFL1C ubi K124    53
#> 2     PSMC1 p Y71    53
#> 3  PSMC1 ubi K258    53
#> 4  SCRIB ubi K105    53
#> 
#> $`54`
#>         PTMnames group
#> 1   PARD3 p S715    54
#> 2 SEC16A p Y1202    54
#> 3   SF3B2 p Y806    54
#> 
#> $`55`
#>         PTMnames group
#> 1  PARD3 p Y1310    55
#> 2 SEC16A p Y1199    55
#> 
#> $`56`
#>        PTMnames group
#> 1   PDZD11 p Y7    56
#> 2   PKP4 p T419    56
#> 3  PKP4 p Y1143    56
#> 4   PSMC1 p Y25    56
#> 5 PTPN11 p Y519    56
#> 6 ZNF185 p Y597    56
#> 
#> $`57`
#>        PTMnames group
#> 1 PIK3R2 p Y577    57
#> 
#> $`58`
#>      PTMnames group
#> 1 PKP4 p T412    58
#> 
#> $`59`
#>         PTMnames group
#> 1    PKP4 p Y482    59
#> 2 PLEKHA6 p Y415    59
#> 3  PTPN11 p S580    59
#> 
#> $`60`
#>         PTMnames group
#> 1 PLEKHA6 p Y729    60
#> 2  PTPN11 p Y515    60
#> 
#> $`61`
#>         PTMnames group
#> 1   SCRIB p Y834    61
#> 2 SCRIB ubi K467    61
#> 3  SPRED1 p Y187    61
#> 
#> $`62`
#>        PTMnames group
#> 1 SEC16A p Y991    62
#> 
#> 
#> $Spearman
#> $`1`
#>           PTMnames group
#> 1    ABLIM1 p Y357     1
#> 2    ABLIM1 p Y373     1
#> 3    DCBLD1 p Y589     1
#> 4    DCBLD1 p Y600     1
#> 5  HSPA1B ubi K126     1
#> 6      MET p Y1003     1
#> 7    PIK3R2 p Y460     1
#> 8     PKP4 p Y1156     1
#> 9     PSMC1 p Y439     1
#> 10    USP15 p Y263     1
#> 
#> $`2`
#>            PTMnames group
#> 1     ABLIM1 p Y383     2
#> 2     ABLIM1 p Y410     2
#> 3     ABLIM1 p Y688     2
#> 4   ABLIM1 ubi K275     2
#> 5      APP ubi K763     2
#> 6       EZR ubi K60     2
#> 7   HNRNPU ubi K352     2
#> 8   HNRNPU ubi K433     2
#> 9    HSPA1B ubi K56     2
#> 10  HSPA1B ubi K597     2
#> 11    MET ubi K1232     2
#> 12   MYH9 ack K1492     2
#> 13    MYH9 ubi K102     2
#> 14     MYH9 ubi K14     2
#> 15    MYH9 ubi K299     2
#> 16    MYH9 ubi K373     2
#> 17   PSMC1 ubi K293     2
#> 18   PSMC1 ubi K430     2
#> 19 PTTG1IP ubi K164     2
#> 20   SEC16A p Y2193     2
#> 21 SLC12A4 ubi K979     2
#> 22 SLC12A4 ubi K988     2
#> 23  SLC9A1 ubi K583     2
#> 24  SLC9A1 ubi K612     2
#> 25  SLC9A1 ubi K625     2
#> 26  SPRED1 ubi K284     2
#> 27    ZDHHC5 p Y497     2
#> 
#> $`3`
#>             PTMnames group
#> 1      ABLIM1 p Y406     3
#> 2      ABLIM1 ubi K8     3
#> 3    ATP5PO ack K158     3
#> 4       ATP5PO p Y35     3
#> 5   CTNNA2 ack K913;     3
#> 6        CTTN p Y289     3
#> 7         CTTN p Y84     3
#> 8      DCBLD1 p Y578     3
#> 9    EEF1A1 ack K395     3
#> 10   EEF1A1 ack K439     3
#> 11       EZR ack K60     3
#> 12    FKBP4 ack K354     3
#> 13       FRS2 p Y392     3
#> 14   HNRNPU ack K343     3
#> 15     HNRNPU p Y257     3
#> 16     HNRNPU p Y260     3
#> 17   HSPA1B ack K348     3
#> 18     HSPA1B p Y41;     3
#> 19       MET p Y1230     3
#> 20     MYH9 ack K545     3
#> 21       MYH9 ack K8     3
#> 22     MYH9 ack K882     3
#> 23         OAT p Y69     3
#> 24      PARD3 p S717     3
#> 25      PARD3 p Y388     3
#> 26     PIK3R2 p Y577     3
#> 27       PKP4 p S406     3
#> 28    PLEKHA6 p S472     3
#> 29    PLEKHA6 p Y729     3
#> 30     PTPN11 p Y515     3
#> 31      PTPRH p Y850     3
#> 32  PTTG1IP ack K169     3
#> 33      RBM14 p Y249     3
#> 34 RNASEH2B ack K284     3
#> 35      SCRIB p Y834     3
#> 36    SCRIB ubi K467     3
#> 37     SEC16A p Y489     3
#> 38     SPRED1 p Y187     3
#> 
#> $`4`
#>        PTMnames group
#> 1 ABLIM1 p Y461     4
#> 2 HNRNPF p Y266     4
#> 3 PHLDB2 p Y301     4
#> 4  PKP4 p Y1143     4
#> 
#> $`5`
#>             PTMnames group
#> 1     ACOX1 ack K437     5
#> 2   EEF1A1 ubi K146;     5
#> 3       EZR ubi K162     5
#> 4    HNRNPU ubi K186     5
#> 5   HSPA1B ubi K451;     5
#> 6   HSPA1B ubi K507;     5
#> 7      MET ack K1248     5
#> 8      MYH9 ack K355     5
#> 9  SLC12A4 ubi K1010     5
#> 10   SLC38A2 ubi K38     5
#> 11  TUBA1B ack K280;     5
#> 12  TUBA1B ack K336;     5
#> 13  TUBA1B ubi K163;     5
#> 
#> $`6`
#>            PTMnames group
#> 1      ACOX1 p Y629     6
#> 2      ACTR5 p Y564     6
#> 3   ATP5PO ack K199     6
#> 4     CTTN ack K161     6
#> 5     CTTN ack K171     6
#> 6      DNM2 p Y125;     6
#> 7   EEF1A1 ack K41;     6
#> 8       EMC7 p S114     6
#> 9       EMC7 p T113     6
#> 10        EZR p Y85     6
#> 11   HNRNPH2 p Y306     6
#> 12  HNRNPH2 ubi K35     6
#> 13  HNRNPU ack K265     6
#> 14  HNRNPU ack K352     6
#> 15    HNRNPU p Y473     6
#> 16 HSPA1B ack K319;     6
#> 17     HSPA9 p Y161     6
#> 18       LCK p Y505     6
#> 19   MYH9 ack K1387     6
#> 20   MYH9 ack K1828     6
#> 21     MYH9 ack K47     6
#> 22   MYH9 ack K682;     6
#> 23   MYH9 ack K821;     6
#> 24    MYO18A p Y700     6
#> 25       NFS1 p Y80     6
#> 26       OAT p Y126     6
#> 27   OXCT1 ack K418     6
#> 28   OXCT1 ack K473     6
#> 29       PHB2 p Y77     6
#> 30    PI4K2A p Y402     6
#> 31      PKP4 p T412     6
#> 32    PTPN11 p Y519     6
#> 33    PTPRH p Y1094     6
#> 34  RASSF8 ubi K176     6
#> 35      RMC1 p Y610     6
#> 36   RNASEH2B p S21     6
#> 37   RNASEH2B p Y23     6
#> 38   RPRD2 ack K671     6
#> 39    SEC16A p Y991     6
#> 40   SF3B2 ack K275     6
#> 41   SUPT16H p Y535     6
#> 42      TNK2 p T857     6
#> 43 TUBA1B ack K370;     6
#> 
#> $`7`
#>          PTMnames group
#> 1    AGFG1 p Y327     7
#> 2 EEF1A1 ack K408     7
#> 
#> $`8`
#>         PTMnames group
#> 1   APH1A p Y256     8
#> 2     APP p Y757     8
#> 3 TMEM106B p Y18     8
#> 
#> $`9`
#>            PTMnames group
#> 1     APH1A ubi K95     9
#> 2       CTTN p S417     9
#> 3     CTTN ubi K181     9
#> 4     CTTN ubi K193     9
#> 5     CTTN ubi K230     9
#> 6     CTTN ubi K272     9
#> 7     DCBLD1 p S619     9
#> 8       DNM2 p Y231     9
#> 9      EZR ubi K450     9
#> 10     EZR ubi K458     9
#> 11   MYH9 ack K1793     9
#> 12   MYH9 ubi K1415     9
#> 13   MYH9 ubi K821;     9
#> 14    PIK3R2 p Y505     9
#> 15      PKP4 p S273     9
#> 16      PKP4 p Y306     9
#> 17    PTPN11 p T547     9
#> 18     RBM14 p Y645     9
#> 19   SEC16A p Y1031     9
#> 20 TUBA1B ubi K163;     9
#> 21 TUBA1B ubi K164;     9
#> 22 TUBA1B ubi K394;     9
#> 23 TUBA1B ubi K401;     9
#> 
#> $`10`
#>            PTMnames group
#> 1     APH1A ubi K96    10
#> 2    ATP5PO ack K60    10
#> 3     CTTN ack K198    10
#> 4    EEF1A1 p T142;    10
#> 5   EEF1A1 ubi K179    10
#> 6   EEF1A1 ubi K392    10
#> 7      EZR ubi K143    10
#> 8      EZR ubi K209    10
#> 9      EZR ubi K211    10
#> 10     EZR ubi K344    10
#> 11     EZR ubi K72;    10
#> 12 HSPA1B ubi K507;    10
#> 13 HSPA1B ubi K512;    10
#> 14      LCK p T395;    10
#> 15      LCK p Y394;    10
#> 16   MYH9 ack K1525    10
#> 17   MYH9 ack K1527    10
#> 18   MYH9 ubi K1638    10
#> 19  NSFL1C ubi K127    10
#> 20    PARD3 p Y1177    10
#> 21  PI4K2A ubi K240    10
#> 22    PKP4 ubi K518    10
#> 23   PRRC2A ack K35    10
#> 24     PTPRK p Y916    10
#> 25    SEC16A p Y813    10
#> 26   SF3B2 ubi K280    10
#> 27   SF3B2 ubi K543    10
#> 28      TNK2 p T517    10
#> 29 TUBA1B ack K336;    10
#> 30   TUBA1B p Y103;    10
#> 31   TUBA1B p Y312;    10
#> 32 TUBA1B ubi K336;    10
#> 33 TUBA1B ubi K338;    10
#> 34 TUBA1B ubi K370;    10
#> 35  TUBA1B ubi K96;    10
#> 36    WEE1 ack K596    10
#> 
#> $`11`
#>             PTMnames group
#> 1         APP p T761    11
#> 2         ATG3 p Y18    11
#> 3    ATP5PO ack K162    11
#> 4    ATP6V0D1 p Y347    11
#> 5  ATP6V0D1 ubi K343    11
#> 6    EEF1A1 ack K392    11
#> 7    EEF1A1 ubi K408    11
#> 8        EZR ubi K83    11
#> 9        FRS2 p Y349    11
#> 10    HNRNPF ack K98    11
#> 11   HNRNPU ubi K524    11
#> 12    HSPA9 ack K159    11
#> 13        LCK p Y394    11
#> 14    MERTK ubi K856    11
#> 15       MET p Y1235    11
#> 16     MET ubi K1190    11
#> 17     MET ubi K1259    11
#> 18        MYH9 p Y11    11
#> 19         MYH9 p Y9    11
#> 20    MYH9 ubi K682;    11
#> 21     PARD3 p Y1127    11
#> 22       PDZD11 p Y7    11
#> 23       PHB2 p Y248    11
#> 24   PI4K2A ubi K336    11
#> 25       PKP4 p Y114    11
#> 26       PKP4 p Y157    11
#> 27    PLEKHA6 p Y498    11
#> 28      PTPRK p Y858    11
#> 29  PTTG1IP ubi K169    11
#> 30      RBM14 p Y226    11
#> 31     SCRIB p Y1360    11
#> 32       SMAGP p Y95    11
#> 33  SUPT16H ack K674    11
#> 34     SYAP1 ack K70    11
#> 35       SYAP1 p Y60    11
#> 36    TMEM106B p Y50    11
#> 37    TMEM192 p Y213    11
#> 38  TUBA1B ack K311;    11
#> 39    TUBA1B p Y103;    11
#> 40    TUBA1B p Y272;    11
#> 41  TUBA1B ubi K112;    11
#> 42  TUBA1B ubi K304;    11
#> 43  TUBA1B ubi K326;    11
#> 44  TUBA1B ubi K336;    11
#> 45  TUBA1B ubi K338;    11
#> 46  TUBA1B ubi K370;    11
#> 47   TUBA1B ubi K60;    11
#> 48   TUBA1B ubi K96;    11
#> 49       WEE1 p S139    11
#> 50       WEE1 p Y132    11
#> 51     ZDHHC5 p S529    11
#> 52     ZDHHC5 p Y533    11
#> 
#> $`12`
#>            PTMnames group
#> 1      APP ubi K751    12
#> 2      ARAP2 p Y473    12
#> 3     CTTN ack K107    12
#> 4       CTTN p S156    12
#> 5       CTTN p S418    12
#> 6       CTTN p S447    12
#> 7       CTTN p Y141    12
#> 8       CTTN p Y154    12
#> 9       CTTN p Y162    12
#> 10      CTTN p Y178    12
#> 11      CTTN p Y199    12
#> 12      CTTN p Y228    12
#> 13      CTTN p Y265    12
#> 14      CTTN p Y334    12
#> 15      CTTN p Y421    12
#> 16      CTTN p Y446    12
#> 17      CTTN p Y453    12
#> 18   EEF1A1 p Y141;    12
#> 19    EEF1A1 p Y29;    12
#> 20  EEF1A1 ubi K172    12
#> 21  EEF1A1 ubi K180    12
#> 22  EEF1A1 ubi K439    12
#> 23  EEF1A1 ubi K44;    12
#> 24     EZR ubi K139    12
#> 25     EZR ubi K79;    12
#> 26      FRS2 p Y306    12
#> 27  HNRNPF ubi K224    12
#> 28  HNRNPU ack K565    12
#> 29  HNRNPU ack K635    12
#> 30  HNRNPU ack K814    12
#> 31  HSPA1B ubi K524    12
#> 32   HSPA9 ack K135    12
#> 33      MET p S1236    12
#> 34      MET p Y1234    12
#> 35     MET ubi K962    12
#> 36      MYH9 p Y754    12
#> 37    MYH9 ubi K651    12
#> 38    MYH9 ubi K833    12
#> 39    MYH9 ubi K856    12
#> 40    MYH9 ubi K860    12
#> 41  NSFL1C ubi K172    12
#> 42    PARD3 p Y1080    12
#> 43     PARD3 p Y489    12
#> 44     PARD3 p Y719    12
#> 45      PKP4 p T477    12
#> 46     PKP4 p Y1094    12
#> 47     PKP4 p Y1115    12
#> 48     PKP4 p Y1168    12
#> 49     PKP4 p Y1172    12
#> 50      PKP4 p Y275    12
#> 51      PKP4 p Y372    12
#> 52      PKP4 p Y415    12
#> 53      PKP4 p Y420    12
#> 54      PKP4 p Y421    12
#> 55      PKP4 p Y425    12
#> 56      PKP4 p Y494    12
#> 57   PLEKHA6 p Y404    12
#> 58   PLEKHA6 p Y492    12
#> 59    PTPN11 p Y279    12
#> 60    PTPN11 p Y546    12
#> 61    PTPN11 p Y584    12
#> 62     PTPN11 p Y62    12
#> 63   PTTG1IP p Y174    12
#> 64   RBM14 ack K164    12
#> 65    SLC38A2 p Y41    12
#> 66  SLC38A2 ubi K59    12
#> 67 TMEM106B ubi K14    12
#> 68 TMEM192 ubi K254    12
#> 69      TNK2 p Y827    12
#> 70      TNK2 p Y859    12
#> 71    ZNF185 p Y581    12
#> 
#> $`13`
#>       PTMnames group
#> 1 ARAP2 p Y334    13
#> 
#> $`14`
#>           PTMnames group
#> 1      ARAP2 p Y77    14
#> 2  ATP5PO ack K192    14
#> 3     FKBP4 p Y411    14
#> 4    NSFL1C p Y167    14
#> 5     NSFL1C p Y95    14
#> 6     PARD3 p S715    14
#> 7     PKP4 p Y1132    14
#> 8      PKP4 p Y369    14
#> 9      PKP4 p Y470    14
#> 10     PKP4 p Y478    14
#> 11   PTPN11 p S595    14
#> 12    PTPRK p Y849    14
#> 13  PTTG1IP p Y165    14
#> 14    RBM14 p Y648    14
#> 15  SEC16A p Y1202    14
#> 16  SEC16A p Y1209    14
#> 17  SEC16A p Y1291    14
#> 18     SGMS2 p Y59    14
#> 19 SLC38A2 ubi K33    14
#> 20   SLC9A1 p Y683    14
#> 21   SPRED1 p Y292    14
#> 22    USP15 p Y274    14
#> 
#> $`15`
#>         PTMnames group
#> 1 CTNNA2 p Y417;    15
#> 2  SLC9A1 p S693    15
#> 
#> $`16`
#>           PTMnames group
#> 1   CTNNAL1 p S538    16
#> 2 EEF1A1 ack K146;    16
#> 3    HNRNPU p S271    16
#> 4    PHLDB2 p Y567    16
#> 5 TUBA1B ubi K311;    16
#> 
#> $`17`
#>           PTMnames group
#> 1   CTNNAL1 p Y534    17
#> 2   CTNNAL1 p Y536    17
#> 3      CTTN p S405    17
#> 4    FKBP4 ubi K28    17
#> 5  HNRNPF ack K167    17
#> 6    HNRNPF p Y210    17
#> 7  HSPA1B ack K246    17
#> 8     LARP1 p Y361    17
#> 9    PARD3 p Y1310    17
#> 10    PRRC2A p Y17    17
#> 11   RASSF8 p Y418    17
#> 12     RBM4 p Y194    17
#> 13   SCRIB p Y1268    17
#> 14  SEC16A p Y1199    17
#> 15   SLC12A4 p Y17    17
#> 16   SLC38A2 p Y28    17
#> 17     TNK2 p Y860    17
#> 18  TUBA1B p Y282;    17
#> 19   ZDHHC5 p Y630    17
#> 20   ZNF185 p Y408    17
#> 21   ZNF185 p Y597    17
#> 
#> $`18`
#>            PTMnames group
#> 1     CTTN ack K124    18
#> 2     CTTN ack K235    18
#> 3     CTTN ack K272    18
#> 4     CTTN ack K309    18
#> 5      CTTN ack K87    18
#> 6       CTTN p Y104    18
#> 7   EEF1A1 ack K172    18
#> 8   EEF1A1 ack K179    18
#> 9  EEF1A1 ack K255;    18
#> 10  EEF1A1 ack K44;    18
#> 11    EEF1A1 p Y177    18
#> 12      EZR p Y270;    18
#> 13 HSPA1B ack K500;    18
#> 14   HSPA9 ack K600    18
#> 15       LCK p Y192    18
#> 16   MYH9 ack K1802    18
#> 17     MYH9 ack K74    18
#> 18     MYH9 p Y1408    18
#> 19     MYH9 p Y1805    18
#> 20  NSFL1C ack K172    18
#> 21      PKP4 p Y443    18
#> 22      PKP4 p Y465    18
#> 23    PTPRH p Y1102    18
#> 24   RBM14 ack K149    18
#> 
#> $`19`
#>       PTMnames group
#> 1  CTTN p T401    19
#> 2 PDZD11 p Y36    19
#> 
#> $`20`
#>            PTMnames group
#> 1       CTTN p T440    20
#> 2   EEF1A1 ubi K395    20
#> 3   HNRNPF ubi K299    20
#> 4  HSPA1B ubi K423;    20
#> 5    HSPA9 ubi K121    20
#> 6     MYH9 ubi K82;    20
#> 7     PSMC1 ubi K69    20
#> 8   PTPRK ubi K1360    20
#> 9     RBM4 ubi K223    20
#> 10   SF3B2 ubi K387    20
#> 11   SF3B2 ubi K394    20
#> 12 TUBA1B ack K304;    20
#> 
#> $`21`
#>         PTMnames group
#> 1    CTTN p Y215    21
#> 2   PKP4 p T1174    21
#> 3 PLEKHA6 p Y471    21
#> 
#> $`22`
#>        PTMnames group
#> 1 DCBLD1 p Y540    22
#> 2  PARD3 p S728    22
#> 3 SPRED1 p Y211    22
#> 
#> $`23`
#>          PTMnames group
#> 1 EEF1A1 ubi K212    23
#> 2     MET p Y1349    23
#> 3     MET p Y1356    23
#> 4  MYH9 ack K1538    23
#> 
#> $`24`
#>          PTMnames group
#> 1 EEF1A1 ubi K273    24
#> 2 HNRNPU ubi K464    24
#> 3  SF3B2 ubi K529    24
#> 
#> $`25`
#>         PTMnames group
#> 1    EZR ack K83    25
#> 2    PKP4 p T419    25
#> 3 TUBA1B p Y357;    25
#> 
#> $`26`
#>          PTMnames group
#> 1   EZR ubi K237;    26
#> 2   MYH9 ubi K760    26
#> 3  MYH9 ubi K821;    26
#> 4 SLC9A1 ubi K690    26
#> 
#> $`27`
#>            PTMnames group
#> 1      EZR ubi K35;    27
#> 2    HNRNPF ubi K98    27
#> 3   HNRNPU ack K464    27
#> 4   HNRNPU ubi K265    27
#> 5   HNRNPU ubi K635    27
#> 6   HSPA1B ubi K325    27
#> 7   HSPA1B ubi K550    27
#> 8     MET ubi K1318    27
#> 9     MYH9 ubi K186    27
#> 10   MYH9 ubi K576;    27
#> 11    MYH9 ubi K850    27
#> 12  PI4K2A ubi K268    27
#> 13   PSMC1 ubi K237    27
#> 14   PSMC1 ubi K396    27
#> 15    PSMC1 ubi K91    27
#> 16    SF3B2 ack K10    27
#> 17  SLC12A4 ubi K97    27
#> 18  SLC9A1 ubi K767    27
#> 19 SUPT16H ack K904    27
#> 20 SUPT16H ubi K596    27
#> 21 SUPT16H ubi K904    27
#> 
#> $`28`
#>        PTMnames group
#> 1 HNRNPF p Y253    28
#> 2  MERTK p Y929    28
#> 3   MET p Y1365    28
#> 
#> $`29`
#>           PTMnames group
#> 1  HNRNPF ubi K167    29
#> 2  HSPA1B ubi K628    29
#> 3 TUBA1B ubi K352;    29
#> 
#> $`30`
#>         PTMnames group
#> 1 HNRNPH2 p Y276    30
#> 2  PHLDB2 p Y162    30
#> 3 SEC16A p Y1380    30
#> 
#> $`31`
#>          PTMnames group
#> 1 HSPA1B ubi K539    31
#> 2   MYH9 ack K651    31
#> 3 NSFL1C ubi K124    31
#> 4     PSMC1 p Y71    31
#> 5  PSMC1 ubi K258    31
#> 6  SCRIB ubi K105    31
#> 
#> $`32`
#>          PTMnames group
#> 1  HSPA9 ack K300    32
#> 2  MYH9 ack K1014    32
#> 3   MYH9 ack K299    32
#> 4  MYH9 ack K565;    32
#> 5   MYH9 ack K613    32
#> 6 PLEKHA6 p Y1044    32
#> 7  PSMC1 ack K258    32
#> 
#> $`33`
#>         PTMnames group
#> 1   MERTK p Y549    33
#> 2 PLEKHA6 p S478    33
#> 
#> $`34`
#>         PTMnames group
#> 1     MET p S990    34
#> 2    PKP4 p S281    34
#> 3    PKP4 p Y390    34
#> 4    PKP4 p Y482    34
#> 5 PLEKHA6 p Y415    34
#> 
#> $`35`
#>         PTMnames group
#> 1  MET ubi K1161    35
#> 2   PARD3 p Y933    35
#> 3 PKP4 ubi K518;    35
#> 4  SPRED1 p Y294    35
#> 
#> $`36`
#>        PTMnames group
#> 1  ORAI1 p Y300    36
#> 2 PIK3R2 p Y464    36
#> 3   PSMC1 p Y25    36
#> 4 PTPN11 p S580    36
#> 
#> $`37`
#>       PTMnames group
#> 1  PKP4 p Y224    37
#> 2 PTPRK p S856    37
#> 3 SF3B2 p Y806    37
#> 4  SGMS2 p Y56    37
#> 5  TJP3 p Y364    37
#> 
#> $`38`
#>         PTMnames group
#> 1    TJP3 p Y361    38
#> 2 ZDHHC5 p Y711;    38
#> 
#> $`39`
#>      PTMnames group
#> 1 TNK2 p Y284    39
#> 
#> 
#> $SED
#> $`1`
#>           PTMnames group
#> 1    ABLIM1 p Y357     1
#> 2    ABLIM1 p Y373     1
#> 3    DCBLD1 p Y589     1
#> 4    DCBLD1 p Y600     1
#> 5  HSPA1B ubi K126     1
#> 6      MET p Y1003     1
#> 7    PIK3R2 p Y460     1
#> 8     PKP4 p Y1156     1
#> 9     PSMC1 p Y439     1
#> 10    USP15 p Y263     1
#> 
#> $`2`
#>           PTMnames group
#> 1    ABLIM1 p Y383     2
#> 2    ABLIM1 p Y410     2
#> 3    DCBLD1 p Y540     2
#> 4    EZR ubi K237;     2
#> 5    MET ubi K1232     2
#> 6    MYH9 ubi K299     2
#> 7    MYH9 ubi K760     2
#> 8   MYH9 ubi K821;     2
#> 9   SEC16A p Y2193     2
#> 10 SLC9A1 ubi K690     2
#> 11   ZDHHC5 p Y497     2
#> 
#> $`3`
#>          PTMnames group
#> 1   ABLIM1 p Y406     3
#> 2 HNRNPU ack K343     3
#> 
#> $`4`
#>        PTMnames group
#> 1 ABLIM1 p Y461     4
#> 2 HNRNPF p Y266     4
#> 3 PHLDB2 p Y301     4
#> 4  PKP4 p Y1143     4
#> 
#> $`5`
#>            PTMnames group
#> 1     ABLIM1 p Y688     5
#> 2   ABLIM1 ubi K275     5
#> 3   EEF1A1 ubi K395     5
#> 4      EZR ubi K35;     5
#> 5       EZR ubi K60     5
#> 6    HNRNPF ubi K98     5
#> 7   HNRNPU ubi K352     5
#> 8   HNRNPU ubi K433     5
#> 9   HNRNPU ubi K635     5
#> 10  HSPA1B ubi K325     5
#> 11  HSPA1B ubi K550     5
#> 12   HSPA1B ubi K56     5
#> 13  HSPA1B ubi K597     5
#> 14    MYH9 ubi K102     5
#> 15     MYH9 ubi K14     5
#> 16    MYH9 ubi K186     5
#> 17    MYH9 ubi K373     5
#> 18   MYH9 ubi K576;     5
#> 19  PI4K2A ubi K268     5
#> 20   PSMC1 ubi K237     5
#> 21   PSMC1 ubi K293     5
#> 22    PSMC1 ubi K69     5
#> 23    RBM4 ubi K223     5
#> 24   SF3B2 ubi K387     5
#> 25   SF3B2 ubi K394     5
#> 26  SLC12A4 ubi K97     5
#> 27 SLC12A4 ubi K979     5
#> 28 SLC12A4 ubi K988     5
#> 29  SLC9A1 ubi K612     5
#> 30  SLC9A1 ubi K625     5
#> 31  SLC9A1 ubi K767     5
#> 32  SPRED1 ubi K284     5
#> 33 SUPT16H ack K904     5
#> 34 SUPT16H ubi K596     5
#> 35 SUPT16H ubi K904     5
#> 
#> $`6`
#>            PTMnames group
#> 1     ABLIM1 ubi K8     6
#> 2      ATP5PO p Y35     6
#> 3  CTNNA2 ack K913;     6
#> 4        CTTN p Y84     6
#> 5     DCBLD1 p Y578     6
#> 6   EEF1A1 ack K395     6
#> 7    FKBP4 ack K354     6
#> 8       FRS2 p Y392     6
#> 9     HNRNPU p Y257     6
#> 10    HNRNPU p Y260     6
#> 11  HSPA1B ack K348     6
#> 12    HSPA1B p Y41;     6
#> 13     MERTK p Y549     6
#> 14      MYH9 ack K8     6
#> 15        OAT p Y69     6
#> 16     PARD3 p Y388     6
#> 17      PKP4 p S406     6
#> 18   PLEKHA6 p S472     6
#> 19   PLEKHA6 p S478     6
#> 20     PTPRH p Y850     6
#> 21 PTTG1IP ack K169     6
#> 22     RBM14 p Y249     6
#> 23    SEC16A p Y489     6
#> 
#> $`7`
#>             PTMnames group
#> 1     ACOX1 ack K437     7
#> 2        CTTN p T440     7
#> 3   EEF1A1 ubi K146;     7
#> 4    EEF1A1 ubi K212     7
#> 5       EZR ubi K162     7
#> 6   HSPA1B ubi K423;     7
#> 7   HSPA1B ubi K451;     7
#> 8   HSPA1B ubi K507;     7
#> 9      MET ack K1248     7
#> 10       MET p Y1349     7
#> 11       MET p Y1356     7
#> 12    MYH9 ack K1538     7
#> 13     MYH9 ack K651     7
#> 14     MYH9 ubi K82;     7
#> 15 SLC12A4 ubi K1010     7
#> 16   SLC38A2 ubi K38     7
#> 17  TUBA1B ubi K163;     7
#> 
#> $`8`
#>            PTMnames group
#> 1      ACOX1 p Y629     8
#> 2      ACTR5 p Y564     8
#> 3      ARAP2 p Y334     8
#> 4     CTTN ack K161     8
#> 5     CTTN ack K171     8
#> 6      DNM2 p Y125;     8
#> 7   EEF1A1 ack K41;     8
#> 8       EMC7 p S114     8
#> 9       EMC7 p T113     8
#> 10        EZR p Y85     8
#> 11   HNRNPH2 p Y306     8
#> 12  HNRNPU ack K265     8
#> 13 HSPA1B ack K319;     8
#> 14     HSPA9 p Y161     8
#> 15       LCK p Y505     8
#> 16     MYH9 ack K47     8
#> 17   MYH9 ack K682;     8
#> 18    MYO18A p Y700     8
#> 19       NFS1 p Y80     8
#> 20       OAT p Y126     8
#> 21   OXCT1 ack K418     8
#> 22   OXCT1 ack K473     8
#> 23       PHB2 p Y77     8
#> 24    PI4K2A p Y402     8
#> 25    PTPN11 p Y519     8
#> 26    PTPRH p Y1094     8
#> 27  RASSF8 ubi K176     8
#> 28      RMC1 p Y610     8
#> 29   RNASEH2B p S21     8
#> 30   RNASEH2B p Y23     8
#> 31   RPRD2 ack K671     8
#> 32    SEC16A p Y991     8
#> 33   SUPT16H p Y535     8
#> 
#> $`9`
#>            PTMnames group
#> 1      AGFG1 p Y327     9
#> 2       CTTN p T401     9
#> 3   EEF1A1 ack K408     9
#> 4       EZR ubi K83     9
#> 5   HNRNPU ubi K524     9
#> 6        LCK p Y394     9
#> 7    MERTK ubi K856     9
#> 8       MET p Y1235     9
#> 9     MET ubi K1259     9
#> 10       MYH9 p Y11     9
#> 11    PARD3 p Y1127     9
#> 12      PKP4 p Y157     9
#> 13     PTPRK p Y858     9
#> 14     RBM14 p Y226     9
#> 15      SMAGP p Y95     9
#> 16 SUPT16H ack K674     9
#> 17 TUBA1B ack K311;     9
#> 18   TUBA1B p Y103;     9
#> 19 TUBA1B ubi K112;     9
#> 20 TUBA1B ubi K326;     9
#> 21 TUBA1B ubi K336;     9
#> 22 TUBA1B ubi K338;     9
#> 23 TUBA1B ubi K370;     9
#> 24  TUBA1B ubi K60;     9
#> 25  TUBA1B ubi K96;     9
#> 26    ZDHHC5 p S529     9
#> 27    ZDHHC5 p Y533     9
#> 
#> $`10`
#>         PTMnames group
#> 1   APH1A p Y256    10
#> 2 TMEM106B p Y18    10
#> 
#> $`11`
#>            PTMnames group
#> 1     APH1A ubi K95    11
#> 2       CTTN p S417    11
#> 3     CTTN ubi K181    11
#> 4     CTTN ubi K193    11
#> 5     CTTN ubi K272    11
#> 6     DCBLD1 p S619    11
#> 7      EZR ubi K450    11
#> 8      EZR ubi K458    11
#> 9    MYH9 ack K1793    11
#> 10   MYH9 ubi K1415    11
#> 11    PIK3R2 p Y505    11
#> 12      PKP4 p S273    11
#> 13      PKP4 p Y306    11
#> 14    PTPN11 p T547    11
#> 15     RBM14 p Y645    11
#> 16   SEC16A p Y1031    11
#> 17 TUBA1B ubi K163;    11
#> 18 TUBA1B ubi K164;    11
#> 19 TUBA1B ubi K394;    11
#> 20 TUBA1B ubi K401;    11
#> 
#> $`12`
#>           PTMnames group
#> 1    APH1A ubi K96    12
#> 2    CTTN ubi K230    12
#> 3 HSPA1B ubi K507;    12
#> 4   MYH9 ubi K821;    12
#> 5  NSFL1C ubi K127    12
#> 
#> $`13`
#>            PTMnames group
#> 1        APP p T761    13
#> 2        ATG3 p Y18    13
#> 3   ATP6V0D1 p Y347    13
#> 4       FRS2 p Y349    13
#> 5     HNRNPF p Y253    13
#> 6      MERTK p Y929    13
#> 7       MET p Y1365    13
#> 8     MET ubi K1190    13
#> 9         MYH9 p Y9    13
#> 10     PDZD11 p Y36    13
#> 11   PLEKHA6 p Y498    13
#> 12 PTTG1IP ubi K169    13
#> 13  SLC38A2 ubi K33    13
#> 14    SYAP1 ack K70    13
#> 15      SYAP1 p Y60    13
#> 16   TMEM106B p Y50    13
#> 17   TMEM192 p Y213    13
#> 18   TUBA1B p Y272;    13
#> 19 TUBA1B ubi K304;    13
#> 20      WEE1 p S139    13
#> 21      WEE1 p Y132    13
#> 
#> $`14`
#>          PTMnames group
#> 1      APP p Y757    14
#> 2 EEF1A1 ack K392    14
#> 3     PHB2 p Y248    14
#> 
#> $`15`
#>            PTMnames group
#> 1      APP ubi K751    15
#> 2   EEF1A1 ubi K172    15
#> 3   EEF1A1 ubi K180    15
#> 4   EEF1A1 ubi K439    15
#> 5   EEF1A1 ubi K44;    15
#> 6      EZR ubi K139    15
#> 7   HNRNPU ack K814    15
#> 8    HSPA9 ack K135    15
#> 9      MET ubi K962    15
#> 10    MYH9 ubi K651    15
#> 11    MYH9 ubi K833    15
#> 12    MYH9 ubi K856    15
#> 13    MYH9 ubi K860    15
#> 14  NSFL1C ubi K172    15
#> 15      PKP4 p Y494    15
#> 16   PLEKHA6 p Y404    15
#> 17  SLC38A2 ubi K59    15
#> 18 TMEM192 ubi K254    15
#> 
#> $`16`
#>          PTMnames group
#> 1    APP ubi K763    16
#> 2 SLC9A1 ubi K583    16
#> 
#> $`17`
#>          PTMnames group
#> 1    ARAP2 p Y473    17
#> 2     CTTN p S156    17
#> 3     CTTN p S447    17
#> 4     CTTN p Y154    17
#> 5     CTTN p Y162    17
#> 6     CTTN p Y178    17
#> 7     CTTN p Y265    17
#> 8     CTTN p Y453    17
#> 9     MET p S1236    17
#> 10   PARD3 p Y489    17
#> 11   PARD3 p Y719    17
#> 12   PKP4 p Y1094    17
#> 13    PKP4 p Y275    17
#> 14    PKP4 p Y420    17
#> 15    PKP4 p Y425    17
#> 16 RBM14 ack K164    17
#> 17  SLC38A2 p Y41    17
#> 18    TNK2 p Y284    17
#> 19  ZNF185 p Y581    17
#> 
#> $`18`
#>           PTMnames group
#> 1      ARAP2 p Y77    18
#> 2  ATP5PO ack K192    18
#> 3   CTNNA2 p Y417;    18
#> 4   CTNNAL1 p Y534    18
#> 5   CTNNAL1 p Y536    18
#> 6      CTTN p S405    18
#> 7     FKBP4 p Y411    18
#> 8    HNRNPF p Y210    18
#> 9  HSPA1B ack K246    18
#> 10    LARP1 p Y361    18
#> 11   NSFL1C p Y167    18
#> 12    NSFL1C p Y95    18
#> 13    ORAI1 p Y300    18
#> 14    PARD3 p S715    18
#> 15   PARD3 p Y1310    18
#> 16   PIK3R2 p Y464    18
#> 17    PKP4 p Y1132    18
#> 18     PKP4 p Y224    18
#> 19     PKP4 p Y369    18
#> 20     PKP4 p Y470    18
#> 21     PKP4 p Y478    18
#> 22    PRRC2A p Y17    18
#> 23   PTPN11 p S595    18
#> 24    PTPRK p S856    18
#> 25    PTPRK p Y849    18
#> 26  PTTG1IP p Y165    18
#> 27   RASSF8 p Y418    18
#> 28    RBM14 p Y648    18
#> 29     RBM4 p Y194    18
#> 30   SCRIB p Y1360    18
#> 31  SEC16A p Y1199    18
#> 32  SEC16A p Y1202    18
#> 33  SEC16A p Y1209    18
#> 34  SEC16A p Y1291    18
#> 35    SF3B2 p Y806    18
#> 36     SGMS2 p Y56    18
#> 37     SGMS2 p Y59    18
#> 38   SLC12A4 p Y17    18
#> 39   SLC38A2 p Y28    18
#> 40   SLC9A1 p S693    18
#> 41   SLC9A1 p Y683    18
#> 42   SPRED1 p Y292    18
#> 43     TJP3 p Y364    18
#> 44     TNK2 p Y860    18
#> 45  TUBA1B p Y282;    18
#> 46  TUBA1B p Y357;    18
#> 47    USP15 p Y274    18
#> 48   ZDHHC5 p Y630    18
#> 49   ZNF185 p Y408    18
#> 50   ZNF185 p Y597    18
#> 
#> $`19`
#>          PTMnames group
#> 1 ATP5PO ack K158    19
#> 
#> $`20`
#>          PTMnames group
#> 1 ATP5PO ack K162    20
#> 2     PDZD11 p Y7    20
#> 
#> $`21`
#>           PTMnames group
#> 1  ATP5PO ack K199    21
#> 2  HNRNPU ack K352    21
#> 3    HNRNPU p Y473    21
#> 4   MYH9 ack K1387    21
#> 5      PKP4 p T412    21
#> 6  PLEKHA6 p Y1044    21
#> 7 TUBA1B ack K370;    21
#> 
#> $`22`
#>            PTMnames group
#> 1    ATP5PO ack K60    22
#> 2     CTTN ack K198    22
#> 3   EEF1A1 ubi K179    22
#> 4   EEF1A1 ubi K392    22
#> 5      EZR ubi K143    22
#> 6      EZR ubi K209    22
#> 7      EZR ubi K211    22
#> 8      EZR ubi K344    22
#> 9      EZR ubi K72;    22
#> 10      LCK p Y394;    22
#> 11   MYH9 ack K1525    22
#> 12   MYH9 ack K1527    22
#> 13   MYH9 ubi K1638    22
#> 14  PI4K2A ubi K240    22
#> 15    PKP4 ubi K518    22
#> 16     PTPRK p Y916    22
#> 17    SEC16A p Y813    22
#> 18   SF3B2 ubi K280    22
#> 19   SF3B2 ubi K543    22
#> 20   TUBA1B p Y103;    22
#> 21 TUBA1B ubi K336;    22
#> 22 TUBA1B ubi K338;    22
#> 23 TUBA1B ubi K370;    22
#> 24  TUBA1B ubi K96;    22
#> 25    WEE1 ack K596    22
#> 
#> $`23`
#>            PTMnames group
#> 1 ATP6V0D1 ubi K343    23
#> 2   EEF1A1 ubi K408    23
#> 3     HNRNPU p S271    23
#> 4    MYH9 ubi K682;    23
#> 5   PI4K2A ubi K336    23
#> 
#> $`24`
#>           PTMnames group
#> 1   CTNNAL1 p S538    24
#> 2 EEF1A1 ack K146;    24
#> 3    PHLDB2 p Y567    24
#> 4 TUBA1B ubi K311;    24
#> 
#> $`25`
#>            PTMnames group
#> 1     CTTN ack K107    25
#> 2     CTTN ack K124    25
#> 3     CTTN ack K235    25
#> 4     CTTN ack K272    25
#> 5     CTTN ack K309    25
#> 6      CTTN ack K87    25
#> 7       CTTN p Y104    25
#> 8   EEF1A1 ack K172    25
#> 9   EEF1A1 ack K179    25
#> 10 EEF1A1 ack K255;    25
#> 11  EEF1A1 ack K44;    25
#> 12    EEF1A1 p Y177    25
#> 13      EZR p Y270;    25
#> 14 HSPA1B ack K500;    25
#> 15   HSPA9 ack K600    25
#> 16       LCK p Y192    25
#> 17   MYH9 ack K1802    25
#> 18     MYH9 ack K74    25
#> 19     MYH9 p Y1408    25
#> 20     MYH9 p Y1805    25
#> 21  NSFL1C ack K172    25
#> 22      PKP4 p Y443    25
#> 23      PKP4 p Y465    25
#> 24    PTPRH p Y1102    25
#> 25   RBM14 ack K149    25
#> 
#> $`26`
#>           PTMnames group
#> 1      CTTN p S418    26
#> 2      CTTN p Y228    26
#> 3     EZR ubi K79;    26
#> 4  HNRNPF ubi K224    26
#> 5  HSPA1B ubi K524    26
#> 6    PARD3 p Y1080    26
#> 7    PTPN11 p Y279    26
#> 8 TMEM106B ubi K14    26
#> 
#> $`27`
#>        PTMnames group
#> 1   CTTN p Y141    27
#> 2   CTTN p Y199    27
#> 3 EEF1A1 p Y29;    27
#> 4   MYH9 p Y754    27
#> 5  PKP4 p Y1172    27
#> 6 PTPN11 p Y546    27
#> 
#> $`28`
#>         PTMnames group
#> 1    CTTN p Y215    28
#> 2   PKP4 p T1174    28
#> 3 PLEKHA6 p Y471    28
#> 4    TJP3 p Y361    28
#> 5 ZDHHC5 p Y711;    28
#> 
#> $`29`
#>         PTMnames group
#> 1    CTTN p Y289    29
#> 2    EZR ack K60    29
#> 3   SCRIB p Y834    29
#> 4 SCRIB ubi K467    29
#> 5  SPRED1 p Y187    29
#> 
#> $`30`
#>           PTMnames group
#> 1      CTTN p Y334    30
#> 2      CTTN p Y421    30
#> 3      CTTN p Y446    30
#> 4   EEF1A1 p Y141;    30
#> 5      FRS2 p Y306    30
#> 6  HNRNPU ack K565    30
#> 7  HNRNPU ack K635    30
#> 8      MET p Y1234    30
#> 9      PKP4 p T477    30
#> 10    PKP4 p Y1115    30
#> 11    PKP4 p Y1168    30
#> 12     PKP4 p Y372    30
#> 13     PKP4 p Y415    30
#> 14     PKP4 p Y421    30
#> 15  PLEKHA6 p Y492    30
#> 16   PTPN11 p Y584    30
#> 17    PTPN11 p Y62    30
#> 18  PTTG1IP p Y174    30
#> 19     TNK2 p Y827    30
#> 20     TNK2 p Y859    30
#> 
#> $`31`
#>           PTMnames group
#> 1      DNM2 p Y231    31
#> 2 HSPA1B ubi K512;    31
#> 
#> $`32`
#>            PTMnames group
#> 1   EEF1A1 ack K439    32
#> 2   HNRNPH2 ubi K35    32
#> 3    MYH9 ack K1828    32
#> 4     MYH9 ack K882    32
#> 5     PIK3R2 p Y577    32
#> 6 RNASEH2B ack K284    32
#> 
#> $`33`
#>         PTMnames group
#> 1 EEF1A1 p T142;    33
#> 2    LCK p T395;    33
#> 3  PARD3 p Y1177    33
#> 4 PRRC2A ack K35    33
#> 
#> $`34`
#>           PTMnames group
#> 1  EEF1A1 ubi K273    34
#> 2  HNRNPU ubi K186    34
#> 3  HNRNPU ubi K464    34
#> 4    MET ubi K1161    34
#> 5     PARD3 p Y933    34
#> 6   PKP4 ubi K518;    34
#> 7 PTTG1IP ubi K164    34
#> 8   SF3B2 ubi K529    34
#> 9    SPRED1 p Y294    34
#> 
#> $`35`
#>          PTMnames group
#> 1     EZR ack K83    35
#> 2   FKBP4 ubi K28    35
#> 3 HNRNPF ack K167    35
#> 4     PKP4 p T419    35
#> 
#> $`36`
#>         PTMnames group
#> 1 HNRNPF ack K98    36
#> 2 HSPA9 ack K159    36
#> 3     MET p S990    36
#> 4    PKP4 p Y482    36
#> 5 PLEKHA6 p Y415    36
#> 6  PTPN11 p S580    36
#> 
#> $`37`
#>           PTMnames group
#> 1  HNRNPF ubi K167    37
#> 2  HSPA1B ubi K628    37
#> 3 TUBA1B ack K280;    37
#> 4 TUBA1B ack K336;    37
#> 5 TUBA1B ubi K352;    37
#> 
#> $`38`
#>           PTMnames group
#> 1  HNRNPF ubi K299    38
#> 2   HSPA9 ubi K121    38
#> 3   MYH9 ack K1492    38
#> 4  PTPRK ubi K1360    38
#> 5 TUBA1B ack K304;    38
#> 
#> $`39`
#>         PTMnames group
#> 1 HNRNPH2 p Y276    39
#> 2  PHLDB2 p Y162    39
#> 3 SEC16A p Y1380    39
#> 
#> $`40`
#>          PTMnames group
#> 1 HNRNPU ack K464    40
#> 2   MET ubi K1318    40
#> 3  PSMC1 ubi K396    40
#> 
#> $`41`
#>          PTMnames group
#> 1 HNRNPU ubi K265    41
#> 2   MYH9 ubi K850    41
#> 3   PSMC1 ubi K91    41
#> 4   SF3B2 ack K10    41
#> 
#> $`42`
#>          PTMnames group
#> 1 HSPA1B ubi K539    42
#> 2 NSFL1C ubi K124    42
#> 3     PSMC1 p Y71    42
#> 4  PSMC1 ubi K258    42
#> 5  SCRIB ubi K105    42
#> 
#> $`43`
#>         PTMnames group
#> 1 HSPA9 ack K300    43
#> 2 MYH9 ack K1014    43
#> 3  MYH9 ack K299    43
#> 4  MYH9 ack K545    43
#> 5 MYH9 ack K565;    43
#> 6  MYH9 ack K613    43
#> 7 PSMC1 ack K258    43
#> 
#> $`44`
#>         PTMnames group
#> 1    MET p Y1230    44
#> 2 MYH9 ack K821;    44
#> 
#> $`45`
#>        PTMnames group
#> 1 MYH9 ack K355    45
#> 
#> $`46`
#>       PTMnames group
#> 1 PARD3 p S717    46
#> 2  TNK2 p T857    46
#> 
#> $`47`
#>         PTMnames group
#> 1   PARD3 p S728    47
#> 2 PSMC1 ubi K430    47
#> 3  SPRED1 p Y211    47
#> 
#> $`48`
#>      PTMnames group
#> 1 PKP4 p S281    48
#> 2 PKP4 p Y390    48
#> 
#> $`49`
#>      PTMnames group
#> 1 PKP4 p Y114    49
#> 
#> $`50`
#>         PTMnames group
#> 1 PLEKHA6 p Y729    50
#> 2  PTPN11 p Y515    50
#> 
#> $`51`
#>      PTMnames group
#> 1 PSMC1 p Y25    51
#> 
#> $`52`
#>        PTMnames group
#> 1 SCRIB p Y1268    52
#> 
#> $`53`
#>         PTMnames group
#> 1 SF3B2 ack K275    53
#> 
#> $`54`
#>           PTMnames group
#> 1      TNK2 p T517    54
#> 2 TUBA1B ack K336;    54
#> 3   TUBA1B p Y312;    54
#> 
#>