Build PPI Network

Uses all inputted database information on protein-protein interactions in order to create a combined PPI-network.

Usage

BuildClusterFilteredNetwork(
  stringdb.edges,
  genemania.edges,
  gene.cccn,
  db.filepaths = c(),
  ppi.network.name = "combined.PPIs",
  cfn.name = "cfn"
)

Arguments

stringdb.edges

Data frame of consisting of the network of interactions from the genes of study pulled from the STRINGdb database

genemania.edges

GeneMANIA network of protein-protein interactions from the genes of study; defaults to NA

gene.cccn

A matrix showing strength of relationships between proteins using the common clusters between the three distance metrics (Euclidean, Spearman, and Combined (SED))

db.filepaths

A vector of paths to the additional ppi network files; defaults to an empty vector

ppi.network.name

Desired name for the output protein-protein interaction network using all entered database input; defaults to ppi.network

cfn.name

Desired name of the output cluster filtered network

ppi.network

A dataframe representing how strongly proteins are known to interact based on previous research contained within provided database networks

Value

A dataframe representing how strongly proteins are known to interact based on previous research contained within provided database networks

A version of ppi.network with only the edges that exist in gene.cccn and have non-zero weights

Examples