Make Correlation Network

Make Correlation Network filters a correlation matrix of PTMs by specific PTMs. It groups the PTM correlation matrix based on the Genes of PTMs. By summing these submatrices, it also produces a gene by gene cocluster correlation network shows strength of relationships between proteins using the common clusters between the three distance metrics.

Usage

MakeCorrelationNetwork(
  common.clusters,
  ptm.correlation.matrix,
  ptm.cccn.name = "ptm.cccn",
  gene.cccn.name = "gene.cccn"
)

Arguments

common.clusters

A list of clusters. Ideally the ones found by MakeClusterList in common_clusters

ptm.correlation.matrix

A data frame showing the correlation between ptms (as the rows and the columns). NAs are placed along the diagonal.

ptm.cccn.name

Desired name of the PTM correlation matrix filtered by specific PTMs from input

gene.cccn.name

Desired name for the cocluster correlation network; defaults to gene.cccn

Value

The list of common clusters between all three distance metrics (Euclidean, Spearman, and SED) and a matrix showing strength of relationships between proteins using the common clusters between the three distance metrics (Euclidean, Spearman, and Combined (SED))

Examples

ex.ptm.cor <- ex.ptm.correlation.matrix
MakeCorrelationNetwork(ex.clusters.list, ex.ptm.cor, 1, "ex.clusters.common", "ex.ptm.cccn", "ex.gene.cccn")
#> Error in MakeCorrelationNetwork(ex.clusters.list, ex.ptm.cor, 1, "ex.clusters.common",     "ex.ptm.cccn", "ex.gene.cccn"): unused arguments ("ex.ptm.cccn", "ex.gene.cccn")
print(ex.clusters.common[c(1, 2, 3)])
#> Error: object 'ex.clusters.common' not found
utils::head(ex.gene.cccn[, c(1,2,3,4,5)])
#>          ABLIM1      MET      PKP4      MYH9    SEC16A
#> ABLIM1  0.00000 40.07949  66.83899  33.44287 16.700350
#> MET    40.07949  0.00000  83.84216  71.98251 19.221441
#> PKP4   66.83899 83.84216   0.00000 202.27512 55.177106
#> MYH9   33.44287 71.98251 202.27512   0.00000 33.703866
#> SEC16A 16.70035 19.22144  55.17711  33.70387  0.000000
#> ZDHHC5 20.71919 22.82946  47.79951  33.60434  7.657868