Pathway Crosstalk Network

Converts Bioplanet pathways from (https://tripod.nih.gov/bioplanet/) into a list of pathways whose elements are the genes in each pathway. Edge weights are either the PTM Cluster Weight or according to the Jaccard Similarity.

Usage

PathwayCrosstalkNetwork(
  common.clusters,
  bioplanet.file = "bioplanet.csv",
  createfile = getwd()
)

Arguments

common.clusters

The list of common clusters between all three distance metrics (Euclidean, Spearman, and SED). Can be made in MakeCorrelationNetwork

bioplanet.file

Either the name of the bioplanet pathway .csv file OR a dataframe. Lines of bioplanet should possess 4 values in the order "PATHWAY_ID","PATHWAY_NAME","GENE_ID","GENE_SYMBOL". Users not well versed in R should only pass in "yourfilename.csv"

createfile

The path of where to create the edgelist file. Defaults to the working directory, if FALSE is provided, a file will not be created.

Value

An edgelist file that is created in the working directory. Contains pathway source-target columns, with edge weights of their jaccard similarity and their Pathway-Pathway Evidence score 1 (pathway.crosstalk.network) Contains pathway source-target columns, along with the interaction type. 2 (PCNedgelist) Contains pathway source-target columns, with edge weights of their jaccard similarity and their Pathway-Pathway Evidence score 3 (pathways.list) All pathways in the bioplanet database as a named list containing string vectors. Each vector is a pathway with strings associated with the genes in that pathway.

Examples

Example_Output <- PathwayCrosstalkNetwork(ex.common.clusters, ex.bioplanet, createfile = FALSE)
#> [1] "Making PCN"
#> [1] "2025-10-19 10:40:08 UTC"
#> [1] "2025-10-19 10:40:08 UTC"
#> [1] Total time: 0.115340232849121
Example_Output[[1]][[3,]]
#> Error in `[[.data.frame`(Example_Output[[1]], 3, ): argument "..2" is missing, with no default
Example_Output[[3]][[1:3]]
#> Error in Example_Output[[3]][[1:3]]: recursive indexing failed at level 2