Pathway Crosstalk Network

Converts Bioplanet pathways from (https://tripod.nih.gov/bioplanet/) into a list of pathways whose elements are the genes in each pathway. Edge weights are either the PTM Cluster Weight or according to the Jaccard Similarity.

Usage

PathwayCrosstalkNetwork(
  common.clusters,
  file = "bioplanet.csv",
  edgelist.name = "PTPedgelist",
  createfile = getwd()
)

Arguments

common.clusters

The list of common clusters between all three distance metrics (Euclidean, Spearman, and SED). Can be made in MakeCorrelationNetwork

file

Either the name of the bioplanet pathway .csv file OR the name of a dataframe. Lines of bioplanet should possess 4 values in the order "PATHWAY_ID","PATHWAY_NAME","GENE_ID","GENE_SYMBOL". Users not well versed in R should only pass in "yourfilename.csv"

edgelist.name

The desired name of the Pathway to Pathway edgelist file created ('.csv' will automatically be added to the end for you); defaults to edgelist. Intended for use in Cytoscape.

createfile

The path of where to create the edgelist file. Defaults to the working directory, if FALSE is provided, a file will not be created.

Value

An edgelist file that is created in the working directory. Contains pathway source-target columns, with edge weights of their jaccard similarity and their Pathway-Pathway Evidence score

Examples

PathwayCrosstalkNetwork(ex.common.clusters, ex.bioplanet, "ex.edgelist", createfile = FALSE)
#> Error in PathwayCrosstalkNetwork(ex.common.clusters, ex.bioplanet, "ex.edgelist",     createfile = FALSE): unused argument (createfile = FALSE)